requirements.txt
torch>=2.0
transformers>=4.40
peft>=0.10
datasets
accelerate
bitsandbytes
trl
pyyaml
anthropic
sentence-transformers
python-dotenv